# perf(git_status): tweak exec flags to omit unnecessary info